Grant SSH access from local /etc/groups
Add a new AllowLocalGroups prop to grant SSH access to groups in
the local /etc/groups file.

The current AllowGroups prop is designed to work with groups in the
accounts provider DB and modified with the web UI.

This new prop is designed to grant access to additional users created by
Anaconda at install time or with config management tools wherever an
accounts provider is not wanted or still not configured.